% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_steplist_creator.R
\name{launch_steplist_creator}
\alias{launch_steplist_creator}
\title{Launch steplist creator \code{shiny} app}
\usage{
launch_steplist_creator()
}
\value{
The \code{launch_steplist_creator} function is used for the side effect of starting the Steplist Creator \code{shiny} app.
}
\description{
Run this function to start the Steplist Creator \code{shiny} app.
}
\examples{
if(interactive()){
launch_steplist_creator()
}
}
