% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_steplist_user.R
\name{remove_all_modules}
\alias{remove_all_modules}
\title{Remove all modules}
\usage{
remove_all_modules(steplist)
}
\arguments{
\item{steplist}{An \code{epicmodel_steplist} or \code{epicmodel_steplist_checked} object.}
}
\value{
An \code{epicmodel_steplist} object with empty data.frame \code{module} and empty strings in variable \code{module_step} in data.frame \code{step}. When
continuing with this steplist, SCC models cannot be inspected by module. If you made any changes, you need to call \code{\link[=check_steplist]{check_steplist()}} again.
}
\description{
Removes all entries in data.frame \code{module} from an \code{epicmodel_steplist} object. Also turns all values of variable \code{module_step} in data.frame
\code{step} from an \code{epicmodel_steplist} to empty strings.
}
\examples{
x <- remove_all_modules(steplist_party)
}
