% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_steplist_user.R
\name{remove_na}
\alias{remove_na}
\title{Removing NA in \code{icc} and \code{outc}}
\usage{
remove_na(steplist)
}
\arguments{
\item{steplist}{An \code{epicmodel_steplist} or \code{epicmodel_steplist_checked} object.}
}
\value{
An \code{epicmodel_steplist} object without entries in data.frame \code{icc}, which contain 'NA' in either \code{id1} or \code{id2} as well as entries in
data.frame \code{outc} that contain 'NA' in \code{id_outc}. If you made any changes, you need to call \code{\link[=check_steplist]{check_steplist()}} again.
}
\description{
Remove any entries that only consist of NA from data.frames \code{icc} (Incompatible Component Causes) and \code{outc} (outcome definition) from an
\code{epicmodel_steplist}.
}
\examples{
x <- remove_na(steplist_party)
}
