% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_steplist_user.R
\name{remove_segment}
\alias{remove_segment}
\title{Remove segments}
\usage{
remove_segment(steplist, id)
}
\arguments{
\item{steplist}{An \code{epicmodel_steplist} or \code{epicmodel_steplist_checked} object.}

\item{id}{A single non-missing element of type character describing the ID of the entry you want deleted.}
}
\value{
An \code{epicmodel_steplist} class object. If you made any changes, you need to call \code{\link[=check_steplist]{check_steplist()}} again.
}
\description{
Removes individual entries from data.frames \code{what}, \code{does}, \code{where}, \code{module}, or \code{icc}.
}
\examples{
steplist_party <- remove_segment(steplist_party, "d4")
}
