% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scc_rain}
\alias{scc_rain}
\title{Rain example SCC model}
\format{
An object of class \code{epicmodel_scc}, which is a list of 10 elements. See \code{\link[=new_scc]{new_scc()}} for the detailed structure of
\code{epicmodel_scc} class objects.
}
\usage{
scc_rain
}
\description{
An example SCC model created from \code{steplist_rain}.
}
\keyword{datasets}
