% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steplist_party}
\alias{steplist_party}
\title{Birthday party example steplist}
\format{
An object of class \code{epicmodel_steplist}, which is a list of 8 data.frames. See \code{\link[=new_steplist]{new_steplist()}} for the detailed structure of
\code{epicmodel_steplist} class objects.
}
\usage{
steplist_party
}
\description{
An example steplist, which contains the steps that tell Clara, under which conditions her birthday party will be a success.
}
\keyword{datasets}
