% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{steplist_rain}
\alias{steplist_rain}
\title{Rain example steplist}
\format{
An object of class \code{epicmodel_steplist}, which is a list of 8 data.frames. See \code{\link[=new_steplist]{new_steplist()}} for the detailed structure of
\code{epicmodel_steplist} class objects.
}
\usage{
steplist_rain
}
\description{
An example steplist, which contains a rain-themed example to illustrate intervention functions.
}
\keyword{datasets}
