% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_steplist_user.R
\name{uncheck_steplist}
\alias{uncheck_steplist}
\title{Unchecking \code{epicmodel_steplist} objects}
\usage{
uncheck_steplist(steplist)
}
\arguments{
\item{steplist}{An \code{epicmodel_steplist} or \code{epicmodel_steplist_checked} object.}
}
\value{
An object of class \code{epicmodel_steplist}.
}
\description{
Putting a checked \code{epicmodel_steplist} back to an unchecked status.
}
\examples{
x <- uncheck_steplist(scc_rain$steplist)
}
